Copy over project from https://github.com/aabiji/cs-homework
The project was kept in a monorepo, but for the sake of record
it would be better to have the project as a separate git repository.
So I just created a new git repository, added cs-homework as a remote
origin, fetched from there, then cherry picked from the commit I
started the project to the HEAD, then removed the remote origin.
Now the git history is preserved and the code is now hosted in
a separate repository.